+++ pilot/pkg/networking/core/v1alpha3/route/route.go
@@ -41,6 +41,7 @@ import (
 	"istio.io/istio/pkg/config/constants"
 	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/labels"
+	"istio.io/istio/pkg/config/validation"
 	"istio.io/istio/pkg/util/gogo"
 	"istio.io/pkg/log"
 )
@@ -282,7 +283,7 @@ func BuildHTTPRoutesForVirtualService(
 				tmp = append(tmp, r)
 				// This is a catch all path. Routes are matched in order, so we will never go beyond this match
 				// As an optimization, we can just drop sending any more routes here.
-				if isCatchAllMatch(match) {
+				if validation.IsCatchAllMatch(match) {
 					// only add the catch all route
 					out = append(out, r)
 					return out, nil
@@ -1132,34 +1133,6 @@ func getHashPolicy(push *model.PushContext, node *model.Proxy, dst *networking.H
 		consistentHash = plsHash
 	}
 	return consistentHashToHashPolicy(consistentHash)
-}
-
-// isCatchAll returns true if HTTPMatchRequest is a catchall match otherwise
-// false. Note - this may not be exactly "catch all" as we don't know the full
-// class of possible inputs As such, this is used only for optimization.
-func isCatchAllMatch(m *networking.HTTPMatchRequest) bool {
-	catchall := false
-	if m.Uri != nil {
-		switch m := m.Uri.MatchType.(type) {
-		case *networking.StringMatch_Prefix:
-			catchall = m.Prefix == "/"
-		case *networking.StringMatch_Regex:
-			catchall = m.Regex == "*"
-		}
-	}
-	// A Match is catch all if and only if it has no match set
-	// and URI has a prefix / or regex *.
-	return catchall &&
-		len(m.Headers) == 0 &&
-		len(m.QueryParams) == 0 &&
-		len(m.SourceLabels) == 0 &&
-		len(m.WithoutHeaders) == 0 &&
-		len(m.Gateways) == 0 &&
-		m.Method == nil &&
-		m.Scheme == nil &&
-		m.Port == 0 &&
-		m.Authority == nil &&
-		m.SourceNamespace == ""
 }
 
 // CombineVHostRoutes semi concatenates Vhost's routes into a single route set.
